change_xml_and_added_user_type_support
