+++ wfe-core/src/main/resources/validators.xml
@@ -21,5 +21,5 @@
     <validator name="containerSize" class="ru.runa.wfe.validation.impl.ContainerSizeValidator"/>
     <validator name="containerElementsRequired" class="ru.runa.wfe.validation.impl.ContainerElementsRequiredValidator"/>
     <validator name="listElementsAreUnique" class="ru.runa.wfe.validation.impl.ListElementsUniqueValidator"/>
-    <validator name="unique-external-storage" class="ru.runa.wfe.validation.impl.UniqueExternalStorageValidator"/>
+    <validator name="unique-storage-user-type-attribute" class="ru.runa.wfe.validation.impl.UniqueStorageUserTypeAttributeValidator"/>
 </validators>
